+++ src/PluginManager/include/PluginManager/AbstractModule.h
@@ -52,12 +52,12 @@ public:
 
     ModuleIntf::RecordIconId getRecordIconId() const noexcept override;
     void setRecording(bool enable) noexcept override;
-    void setPauseMode(SkyConnectIntf::PauseMode pauseMode) noexcept override;
+    void setPaused(bool enable) noexcept override;
     void setPlaying(bool enable) noexcept override;
 
 protected:
     virtual void onStartRecording() noexcept;
-    virtual void onPauseMode(SkyConnectIntf::PauseMode pauseMode) noexcept;
+    virtual void onPause(bool enable) noexcept;
     virtual void onStartReplay() noexcept;
 
     FlightService &getFlightService() const noexcept;