+++ src/PluginManager/include/PluginManager/ModuleIntf.h
@@ -53,7 +53,7 @@ public:
     virtual RecordIconId getRecordIconId() const noexcept = 0;
 
     virtual void setRecording(bool enable) noexcept = 0;
-    virtual void setPauseMode(SkyConnectIntf::PauseMode pauseMode) noexcept = 0;
+    virtual void setPaused(bool enable) noexcept = 0;
     virtual void setPlaying(bool enable) noexcept = 0;
 
 protected: