+++ src/Plugins/Connect/PathCreator/src/PathCreatorPlugin.h
@@ -63,7 +63,7 @@ protected:
     void onStopRecording() noexcept override;
 
     bool onStartReplay(std::int64_t currentTimestamp) noexcept override;
-    void onReplayPaused(PauseMode pauseMode) noexcept override;
+    void onReplayPaused(bool enable) noexcept override;
     void onStopReplay() noexcept override;
 
     void onSeek(std::int64_t currentTimestamp) noexcept override;