+++ src/UserInterface/src/MainWindow.cpp
@@ -1091,9 +1091,9 @@ void MainWindow::onPositionSliderPressed() noexcept
 {
     SkyConnectManager &skyConnectManager = SkyConnectManager::getInstance();
     d->previousState = skyConnectManager.getState();
-    if (d->previousState == Connect::State::Replay) {
+    if (skyConnectManager.isInReplayState()) {
         // Pause the replay while sliding the position slider
-        skyConnectManager.setPauseMode(SkyConnectIntf::PauseMode::PauseDuringSeek);
+        skyConnectManager.setPaused(true);
     }
 }
 
@@ -1114,7 +1114,9 @@ void MainWindow::onPositionSliderReleased() noexcept
 {
     SkyConnectManager &skyConnectManager = SkyConnectManager::getInstance();
     if (d->previousState == Connect::State::Replay) {
-        skyConnectManager.setPauseMode(SkyConnectIntf::PauseMode::Resume);
+        skyConnectManager.setPaused(false);
+    } else if (d->previousState == Connect::State::ReplayPaused) {
+        skyConnectManager.setPaused(true);
     }
 }
 
@@ -1620,8 +1622,7 @@ void MainWindow::toggleRecord(bool enable) noexcept
 
 void MainWindow::togglePause(bool enable) noexcept
 {
-    SkyConnectIntf::PauseMode pauseMode = enable ? SkyConnectIntf::PauseMode::Pause : SkyConnectIntf::PauseMode::Resume;
-    d->moduleManager->setPauseMode(pauseMode);
+    d->moduleManager->setPaused(enable);
 }
 
 void MainWindow::togglePlay(bool enable) noexcept
@@ -1670,11 +1671,11 @@ void MainWindow::onFlightRestored() noexcept
     skyConnectManager.skipToBegin();
     if (skyConnectManager.isConnected()) {
         // Make sure we are unpaused...
-        d->moduleManager->setPauseMode(SkyConnectIntf::PauseMode::Resume);
+        d->moduleManager->setPaused(false);
         // ... play the first frame (which will "move" to the new location)...
         d->moduleManager->setPlaying(true);
         // ... and pause again (such that the new scenery can be loaded)
-        d->moduleManager->setPauseMode(SkyConnectIntf::PauseMode::Pause);
+        d->moduleManager->setPaused(true);
     }
 }
 
@@ -1695,7 +1696,7 @@ void MainWindow::onFlightImport(QAction *action) noexcept
         skyConnectManager.skipToBegin();
         if (skyConnectManager.isConnected()) {
             d->moduleManager->setPlaying(true);
-            d->moduleManager->setPauseMode(SkyConnectIntf::PauseMode::Pause);
+            d->moduleManager->setPaused(true);
         }
     }
 }